I have updated MOP.pm to from doing a great deal of "::DISPATCH( $::(class), 'new', (args) )" to using the make_class "macro".  This has greatly improved readability.
I compared the test results before and after the addition of this code and they appear to be the same.

One note to be careful about when making this transition is that a number of "$::Methods"s were installed originally in this manner.

... 'true', ::DISPATCH( $::Method, 'new', sub { ::DISPATCH( $::Bit, 'new', 1 ); } ) ...
is converted to 
$::(something) = make_class ( ...
   methods => {
       'true' => sub { ::DISPATCH( $::Bit, 'new', 1 ); }
   }
);

The "intervening" ::DISPATCH( $::Method, 'new', sub { ... } ) is done for you in make_class

   


git-svn-id: http://svn.pugscode.org/pugs@18756 c213334d-75ef-0310-aa23-eaa082d1ae64