[kp6]
I have serious doubts about this commit (fglock - ping!).  fglock directed me towards the APPLY / Signature and so
on, however,  it appears that $::Str just needed a "say" method added to it.

The doubts are mostly because what fglock said does not even remotely apply to the fix.

http://irclog.perlgeek.de/perl6/2007-11-27#i_151982

dlocaus @ #perl6 irc.freenode.net


git-svn-id: http://svn.pugscode.org/pugs@18945 c213334d-75ef-0310-aa23-eaa082d1ae64