This is a very drastic revision (hopefully this won't turn into a revert war ;)
Here's the changes in summary:

removed all references to ...

Locales, including eras, which come from a locale - this is a vast and complicated domain

Alternate calendars - also vast and complicated

String parsing of any sort - ditto, see the pattern here? ;)

Format specifiers - this could come from locales (CLDR specifies this)
or strftime, but again, it's more complicated than is needed

Comparing dates or times to durations - this just doesn't make
sense. Is 2009-02-23 greater or less than 5 days?

Any sort of date or time math

Added iso8601 output for every role, and made that the
stringification. ISO8601 is unambiguous world-wide, easy to read, and
easy to output.

Renamed Temporal::Instant to Temporal::DateTime

Got rid of Temporal::Subsecond and just made Temporal::Time allow for
sub-second resolutions. Not sure if this is best done with an
$.attosecond attribute or as a decimal number.

Renamed Temporal::Timezone to Temporal::TimeZone::Observance. The
latter is a simple thing which represents the offset, isdst flag, and
short name for a given local time. This information should be
available on all supported platforms. TimeZones themselves are
complicated and very much platform-dependent. Better to leave this as
a separate CPAN6 distro.

Got rid of all mutating operators on everything. The built-ins should
be immutable for simplicity.

Added numification overloading for Temporal::DateTime, which gives us
comparison for free.


git-svn-id: http://svn.pugscode.org/pugs@25445 c213334d-75ef-0310-aa23-eaa082d1ae64