[S32/IO] changed API for C<.read> and C<.write>
The API looked unnatural and non-Perly, so I changed it to something that made
sense. Rather than sending in an Int to indicate how much to write, just let
the length of the Buf determine that. (If you want to send a prefix of a Buf,
just make a slice -- a more general operation.)

Similarly, rather than having to send in an existing Buf for overwriting in
the read method, just return a new Buf. Don't need to return an Int; if we
are truly interested in the number of bytes read, we can do $buf.elems.

git-svn-id: http://svn.pugscode.org/pugs@31543 c213334d-75ef-0310-aa23-eaa082d1ae64