Reword the text of perlcabal.org to diassociate "synopsis" and "spec".
I've long felt that since the synopses contain quite a lot of speculation,
treating them as "formal specification" is not correct.  And Synopsis 1
has long indicated that Perl 6 compliance is defined by the official
test suite, and not by compliance with the synopses or an implementation.

So, in the spirit of "forgiveness" > "permission" this commit is a
somewhat heavy-handed declaration that "synopses are not the formal
specification".  If there's a consensus that this is taking things too
far, feel free to revise or revert this commit until a consensus can
be reached.