Adding Roman Numerals
I'll just reprise a blog post because I'm going to be somewhat pressed for time. I *really* want to do Babylonian but that'd mean rewriting the Perl 6 core a bit. The other reason is that the cuneiform numbers aren't exactly common in Unicode fonts, so it won't play as well. But I *will* do it, damnit.